+++ org.dsource.ddt.ide.core/src/mmrnmhrm/core/CoreTaskAgent.java
@@ -24,7 +24,7 @@ public class CoreTaskAgent extends ExecutorTaskAgent {
 		// Log unexpected exceptions. This is important for two reasons:
 		// 1: Give some user feedback an internal error ocurred, even if is for the log only.
 		// 2: For tests to be able to determine if exceptions that should fail the test have occurred.
-		LangCore.logError(throwable, "Unhandled exception in CoreExecutor: " + getName());
+		LangCore.logError("Unhandled exception in CoreExecutor: " + getName(), throwable);
 	}
 	
 }