+++ org.dsource.ddt.ide.core/src/mmrnmhrm/core/codeassist/DeeProjectModuleResolver.java
@@ -6,15 +6,20 @@ import static melnorme.utilbox.core.CoreUtil.areEqualArrays;
 import java.util.ArrayList;
 import java.util.List;
 
+import melnorme.lang.ide.core.utils.EclipseUtils;
 import melnorme.utilbox.misc.ArrayUtil;
 import melnorme.utilbox.misc.StringUtil;
 import mmrnmhrm.core.DLTKModelUtils;
+import mmrnmhrm.core.DeeCore;
 import mmrnmhrm.core.parser.DeeModuleParsingUtil;
 
-import org.dsource.ddt.ide.core.DeeLanguageToolkit;
+import org.dsource.ddt.ide.core.DeeNature;
+import org.eclipse.core.resources.IProject;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IPath;
 import org.eclipse.core.runtime.Path;
+import org.eclipse.dltk.core.DLTKCore;
+import org.eclipse.dltk.core.IBuildpathEntry;
 import org.eclipse.dltk.core.IModelElement;
 import org.eclipse.dltk.core.IProjectFragment;
 import org.eclipse.dltk.core.IScriptFolder;
@@ -79,70 +84,154 @@ public class DeeProjectModuleResolver extends CommonModuleResolver{
 		if(moduleUnit == null)
 			return null;
 		
-		Module module = DeeModuleParsingUtil.getParsedDeeModule(moduleUnit);
-		return module;
+		return DeeModuleParsingUtil.getParsedDeeModule(moduleUnit);
 	}
 	
 	public ISourceModule findModuleUnit(IScriptProject deeProject, String[] packages, String moduleName) 
 			throws ModelException {
 		IPath packagesPath = new Path(StringUtil.collToString(packages, "/"));
 		
-		return findModuleForImport(deeProject, moduleName, packagesPath);
+		return new ModuleImportTargetFinder(moduleName, packagesPath).findImportTarget(deeProject);
 	}
 	
-	protected ISourceModule findModuleForImport(IScriptProject deeProject, String moduleName, IPath packagesPath)
-			throws ModelException {
-		if(deeProject.exists() == false || !isDeeProject(deeProject))
-			return null;
+	protected static class ImportPathVisitor {
 		
-		for (IProjectFragment srcFolder : deeProject.getProjectFragments()) {
-			ISourceModule srcModule = findModuleForImport(moduleName, packagesPath, srcFolder);
-			if(srcModule != null) {
-				return srcModule;
+		protected static boolean isAccessible(IScriptProject scriptProject) {
+			return DeeNature.isAcessible(scriptProject.getProject(), false);
+		}
+		
+		protected void iteratonFullImportPath(IScriptProject deeProject) throws ModelException {
+			if(!isAccessible(deeProject))
+				return;
+			
+			for (IProjectFragment srcFolder : deeProject.getProjectFragments()) {
+				boolean stop = visitSourceContainer(srcFolder);
+				if(stop) {
+					return;
+				}
+			}
+			
+			IBuildpathEntry[] resolvedBuildpath = deeProject.getResolvedBuildpath(true);
+			for (IBuildpathEntry bpEntry : resolvedBuildpath) {
+				if(bpEntry.getEntryKind() == IBuildpathEntry.BPE_PROJECT) {
+					IPath path = bpEntry.getPath();
+					if(path.segmentCount() != 1) {
+						DeeCore.logError("Invalid path in project BP entry: " + path);
+						continue;
+					}
+					String projectName = path.segment(0);
+					IScriptProject depProject = getDeeScriptProject(projectName);
+					if(depProject == null) {
+						continue;
+					}
+					IProjectFragment[] projectFragments = depProject.getProjectFragments();
+					for (IProjectFragment projectFragment : projectFragments) {
+						if(projectFragment.isExternal()) {
+							continue;
+						}
+						
+						boolean stop = visitSourceContainer(projectFragment);
+						if(stop) {
+							return;
+						}
+					}
+				}
 			}
 		}
 		
-		return null;
+		@SuppressWarnings("unused")
+		protected boolean visitSourceContainer(IProjectFragment srcContainer) throws ModelException {
+			return false;
+		}
+		
+		protected IScriptProject getDeeScriptProject(String projectName) {
+			IProject project = EclipseUtils.getWorkspaceRoot().getProject(projectName);
+			try {
+				if(DeeNature.isAcessible(project)) {
+					return DLTKCore.create(project);
+				}
+			} catch (CoreException e) {
+			}
+			return null;
+		}
+		
 	}
 	
-	protected ISourceModule findModuleForImport(String moduleName, IPath packagesPath, IProjectFragment srcFolder) {
-		IScriptFolder scriptFolder = srcFolder.getScriptFolder(packagesPath);
+	protected static class ModuleImportTargetFinder extends ImportPathVisitor {
 		
-		if(scriptFolder.exists()) {
+		protected final String moduleName;
+		protected final IPath packagesPath;
+		
+		protected ISourceModule foundSourceModule;
+		
+		public ModuleImportTargetFinder(String moduleName, IPath packagesPath) {
+			this.moduleName = moduleName;
+			this.packagesPath = packagesPath;
+		}
+		
+		public ISourceModule findImportTarget(IScriptProject deeProject) throws ModelException {
+			iteratonFullImportPath(deeProject);
+			return foundSourceModule;
+		}
+		
+		@Override
+		protected boolean visitSourceContainer(IProjectFragment srcFolder) {
+			IScriptFolder scriptFolder = srcFolder.getScriptFolder(packagesPath);
+			
+			if(scriptFolder.exists()) {
+				foundSourceModule = getExistingSourceModule(scriptFolder, moduleName);
+				if(foundSourceModule != null) {
+					return true;
+				}
+			}
+			// search for package.d
+			scriptFolder = srcFolder.getScriptFolder(packagesPath.append(moduleName));
+			if(scriptFolder.exists()) {
+				foundSourceModule = getExistingSourceModule(scriptFolder, "package");
+				if(foundSourceModule != null) {
+					return true;
+				}
+			}
+			return false;
+		}
+		
+		protected static ISourceModule getExistingSourceModule(IScriptFolder scriptFolder, String moduleName) {
 			for (String validExtension : DeeNamingRules.VALID_EXTENSIONS) {
 				ISourceModule sourceModule = scriptFolder.getSourceModule(moduleName + validExtension);
 				if(DLTKModelUtils.exists(sourceModule)) {
 					return sourceModule;
 				}
-				// search for package.d
-				ISourceModule srcModule = findModuleForImport("package", packagesPath.append(moduleName), srcFolder);
-				if(srcModule != null) {
-					return srcModule;
-				}
 			}
+			return null;
 		}
-		return null;
-	}
-	
-	protected static boolean isDeeProject(IScriptProject scriptProject) {
-		return scriptProject.getLanguageToolkit().getNatureId().equals(DeeLanguageToolkit.NATURE_ID);
 	}
 	
 	@Override
 	public String[] findModules_do(String fqNamePrefix) throws ModelException {
-		return findModules(fqNamePrefix, this.scriptProject);
+		return new ImportsModuleFinder(fqNamePrefix).findModules(scriptProject);
 	}
 	
-	protected static String[] findModules(String fqNamePrefix, IScriptProject scriptProject) throws ModelException {
-		List<String> matchedModulesFQName = new ArrayList<String>();
+	protected static class ImportsModuleFinder extends ImportPathVisitor {
 		
-		for (IProjectFragment srcFolder : scriptProject.getProjectFragments()) {
-			
-			for (IModelElement pkgFragElem : srcFolder.getChildren()) {
+		protected final String fqNamePrefix;
+		protected List<String> matchedModulesFQName = new ArrayList<String>();
+		
+		public ImportsModuleFinder(String fqNamePrefix) {
+			this.fqNamePrefix = fqNamePrefix;
+		}
+		
+		protected String[] findModules(IScriptProject scriptProject) throws ModelException {
+			iteratonFullImportPath(scriptProject);
+			return ArrayUtil.createFrom(matchedModulesFQName, String.class);
+		}
+		
+		@Override
+		protected boolean visitSourceContainer(IProjectFragment srcContainer) throws ModelException {
+			for (IModelElement pkgFragElem : srcContainer.getChildren()) {
 				IScriptFolder pkgFrag = (IScriptFolder) pkgFragElem;
 				
 				String packagePath= pkgFrag.getElementName();
-				/*BUG here with package.d*/
+				
 				if(!DeeNamingRules.isValidPackagePathName(packagePath))
 					continue;
 				
@@ -160,9 +249,9 @@ public class DeeProjectModuleResolver extends CommonModuleResolver{
 					}
 				}
 			}
+			return false;
 		}
 		
-		return ArrayUtil.createFrom(matchedModulesFQName, String.class);
 	}
 	
 }